Extract TestSceneFocus
